Time: 63 ms (91.38%), Space: 9.3 MB (67.38%) - LeetHub

+++ 823-binary-trees-with-factors/823-binary-trees-with-factors.cpp
@@ -1,12 +1,12 @@
 using ll = long long;
-const ll mod = 1e9+7;
+const int mod = 1e9+7;
 class Solution {
 public:
     int numFactoredBinaryTrees(vector<int>& a) {
         int n = a.size();
         sort(a.begin(), a.end());
         vector<ll> dp(n,1);
-        unordered_map<ll,ll> mp;
+        unordered_map<int,int> mp;
         for (int i=0; i<n; i++) mp[a[i]] = i;
         for (int i=0; i<n; i++) {
           for (int j=0; j<i; j++) {
@@ -18,7 +18,7 @@ public:
               }
           }
         }
-        ll ans = 0;
+        int ans = 0;
         for (int i=0; i<n; i++) {
           ans += dp[i];
           ans %= mod;
